requirements.txt.txt
fastapi
uvicorn
psycopg2-binary
jinja2
pandas
openpyxl
